requirements.txt
groq
cerebras
python-dotenv
playwright
pillow